backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
boto3==1.34.0
# langchain removed - using simple text splitter instead (Python 3.14 compatibility)
# langchain-postgres==0.0.6  # Not needed - using pgvector directly
pgvector==0.2.5
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
django-environ==0.11.2
numpy>=1.24.0
google-generativeai>=0.3.0

